feat: 포지션 뱃지 클릭 토스트 훅 작성
