+++ src/hooks/usePositionToast.tsx
@@ -1,0 +1,38 @@
+import toast from 'react-hot-toast';
+
+import { Flex } from '@components/shared/Flex';
+import { Text } from '@components/shared/Text';
+
+import { usePositionsQuery } from '@hooks/queries/usePositionsQuery';
+
+import { Position } from '@type/models/Position';
+
+export const usePositionToast = () => {
+  const { data: positions } = usePositionsQuery();
+
+  const handleClickPosition = (myPosition: Position) => {
+    const positionInfo = positions.find(
+      (position) => position.acronym === myPosition
+    );
+
+    if (!positionInfo) {
+      return;
+    }
+
+    toast(
+      () => (
+        <Flex align="center">
+          <Flex direction="column" align="center">
+            <Text size={18} weight={700}>
+              {positionInfo.name}
+            </Text>
+            <Text>{positionInfo.description}</Text>
+          </Flex>
+        </Flex>
+      ),
+      { position: 'bottom-center' }
+    );
+  };
+
+  return { handleClickPosition };
+};
