+++ src/pages/ProfilePage/Profile.tsx
@@ -3,22 +3,21 @@ import toast from 'react-hot-toast';
 import { useNavigate } from 'react-router-dom';
 
 import { Avatar } from '@components/Avatar';
-import { Modal } from '@components/Modal';
 import { Button } from '@components/shared/Button';
 import { Flex } from '@components/shared/Flex';
 import { Image } from '@components/shared/Image';
 import { Text } from '@components/shared/Text';
 
 import { useMemberProfileQuery } from '@hooks/queries/useMemberProfileQuery';
-import { usePositionsQuery } from '@hooks/queries/usePositionsQuery';
 import { useChatOnButtonClick } from '@hooks/useChatOnButtonClick';
+import { usePositionToast } from '@hooks/usePositionToast';
 
 import { theme } from '@styles/theme';
 
 import { useLoginInfoStore } from '@stores/loginInfo.store';
 
 import { Member } from '@type/models';
-import { Position, PositionInfo } from '@type/models/Position';
+import { PositionInfo } from '@type/models/Position';
 
 import { PATH_NAME } from '@consts/pathName';
 
@@ -26,7 +25,7 @@ import Social from '@assets/follow.svg?react';
 import HandHeart from '@assets/handHeart.svg?react';
 import Heart from '@assets/heart.svg?react';
 
-import { ModalItem, PointerFlex } from './ProfilePage.style';
+import { PointerFlex } from './ProfilePage.style';
 import {
   ColoredSvgWrapper,
   CrewGroup,
@@ -62,12 +61,9 @@ export const Profile = ({ memberId }: { memberId: Member['id'] }) => {
   const navigate = useNavigate();
 
   const { data: profile } = useMemberProfileQuery({ memberId });
-  const { data: positions } = usePositionsQuery();
 
   const [isHeartClicked, setIsHeartClicked] = useState(false);
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [clickedPositionInfo, setClickedPositionInfo] =
-    useState<PositionInfo | null>(null);
+  useState<PositionInfo | null>(null);
 
   const handleClickHeart = () => {
     setIsHeartClicked((prev: boolean) => !prev);
@@ -84,23 +80,11 @@ export const Profile = ({ memberId }: { memberId: Member['id'] }) => {
     myId,
   });
 
-  const handleClickPosition = (myPosition: Position) => {
-    const positionInfo = positions.find(
-      (position) => position.acronym === myPosition
-    );
-
-    if (!positionInfo) {
-      return;
-    }
-
-    setClickedPositionInfo(positionInfo);
-
-    setIsModalOpen(true);
-  };
-
   const handleClickCrew = (id: Member['id']) => {
     moveToPage(PATH_NAME.GET_CREWS_PATH(String(id)));
   };
+
+  const { handleClickPosition } = usePositionToast();
 
   return (
     <Main>
@@ -190,18 +174,6 @@ export const Profile = ({ memberId }: { memberId: Member['id'] }) => {
           <Introduce>{profile.introduction}</Introduce>
         </ProfileField>
       </FlexItem>
-      <Modal isOpen={isModalOpen} close={() => setIsModalOpen(false)}>
-        {clickedPositionInfo && (
-          <Modal.Content>
-            <ModalItem direction="column" align="center" gap={8}>
-              <Text size={24} weight={700}>
-                {clickedPositionInfo.name}
-              </Text>
-              <Text>{clickedPositionInfo.description}</Text>
-            </ModalItem>
-          </Modal.Content>
-        )}
-      </Modal>
     </Main>
   );
 };
